Logic for resturant.
I've added a menu and an admin who can addor delete items from the menu.
There's an employee who can add or delete items from an order. An order
is stored in a list after being confirmed. The total price for an order
is totaled. The only test made so far doesn't work.